+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Refactor `Impl Future` parameters into `impl IntoFuture`. 
 * Implement `IntoFuture for ResponseVar<T>`.
 
 # 0.11.1